# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"BTC Forecast API","version":"0.1.0"},"paths":{"/api/v1/ingest/metadata":{"post":{"tags":["ingestion"],"summary":"Trigger Ingest Metadata","operationId":"trigger_ingest_metadata_api_v1_ingest_metadata_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/ingest/candles":{"post":{"tags":["ingestion"],"summary":"Trigger Ingest Candles","operationId":"trigger_ingest_candles_api_v1_ingest_candles_post","parameters":[{"name":"candles_csv","in":"query","required":false,"schema":{"type":"string","default":"xbx_coindesk_ohlcv_1d_clean.csv","title":"Candles Csv"}},{"name":"csv_path","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Csv Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ingest/fgi":{"post":{"tags":["ingestion"],"summary":"Trigger Ingest Fgi","operationId":"trigger_ingest_fgi_api_v1_ingest_fgi_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/ingest/all":{"post":{"tags":["ingestion"],"summary":"Trigger Ingest All","operationId":"trigger_ingest_all_api_v1_ingest_all_post","parameters":[{"name":"candles_csv","in":"query","required":false,"schema":{"type":"string","default":"xbx_coindesk_ohlcv_1d_clean.csv","title":"Candles Csv"}},{"name":"csv_path","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Csv Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ingest/sync":{"post":{"tags":["ingestion"],"summary":"Trigger Sync","operationId":"trigger_sync_api_v1_ingest_sync_post","parameters":[{"name":"candles_csv","in":"query","required":false,"schema":{"type":"string","default":"xbx_coindesk_ohlcv_1d_clean.csv","title":"Candles Csv"}},{"name":"csv_path","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Csv Path"}},{"name":"lookback_days","in":"query","required":false,"schema":{"type":"integer","default":3,"title":"Lookback Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ingest/sync/background":{"post":{"tags":["ingestion"],"summary":"Trigger Sync Background","operationId":"trigger_sync_background_api_v1_ingest_sync_background_post","parameters":[{"name":"candles_csv","in":"query","required":false,"schema":{"type":"string","default":"xbx_coindesk_ohlcv_1d_clean.csv","title":"Candles Csv"}},{"name":"csv_path","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Csv Path"}},{"name":"lookback_days","in":"query","required":false,"schema":{"type":"integer","default":3,"title":"Lookback Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/market/candles":{"get":{"tags":["market"],"summary":"Get Candles","operationId":"get_candles_api_v1_market_candles_get","parameters":[{"name":"symbol","in":"query","required":true,"schema":{"type":"string","description":"Symbol identifier e.g. XBX-USD","title":"Symbol"},"description":"Symbol identifier e.g. XBX-USD"},{"name":"interval","in":"query","required":false,"schema":{"type":"string","default":"1d","title":"Interval"}},{"name":"start","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start"}},{"name":"end","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":1000,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CandleSchema"},"title":"Response Get Candles Api V1 Market Candles Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/market/fgi":{"get":{"tags":["market"],"summary":"Get Fgi","operationId":"get_fgi_api_v1_market_fgi_get","parameters":[{"name":"start","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start"}},{"name":"end","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":2000,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FgiDailySchema"},"title":"Response Get Fgi Api V1 Market Fgi Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/market/features":{"get":{"tags":["market"],"summary":"Get Features","operationId":"get_features_api_v1_market_features_get","parameters":[{"name":"symbol","in":"query","required":true,"schema":{"type":"string","description":"Symbol identifier e.g. XBX-USD","title":"Symbol"},"description":"Symbol identifier e.g. XBX-USD"},{"name":"interval","in":"query","required":false,"schema":{"type":"string","default":"1d","title":"Interval"}},{"name":"feature_set","in":"query","required":false,"schema":{"type":"string","default":"full","title":"Feature Set"}},{"name":"start","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start"}},{"name":"end","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":1000,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FeatureSchema"},"title":"Response Get Features Api V1 Market Features Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/train":{"post":{"tags":["ml"],"summary":"Train","operationId":"train_api_v1_train_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/predict":{"post":{"tags":["ml"],"summary":"Predict","operationId":"predict_api_v1_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health/live":{"get":{"tags":["health"],"summary":"Health Live","operationId":"health_live_health_live_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/ready":{"get":{"tags":["health"],"summary":"Health Ready","operationId":"health_ready_health_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CandleSchema":{"properties":{"open_time":{"type":"string","format":"date-time","title":"Open Time"},"open":{"type":"number","title":"Open"},"high":{"type":"number","title":"High"},"low":{"type":"number","title":"Low"},"close":{"type":"number","title":"Close"},"volume":{"type":"number","title":"Volume"}},"type":"object","required":["open_time","open","high","low","close","volume"],"title":"CandleSchema"},"FeatureSchema":{"properties":{"open_time":{"type":"string","format":"date-time","title":"Open Time"},"values":{"additionalProperties":true,"type":"object","title":"Values"}},"type":"object","required":["open_time","values"],"title":"FeatureSchema"},"FgiDailySchema":{"properties":{"open_time":{"type":"string","format":"date-time","title":"Open Time"},"fgi":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Fgi"},"fgi_norm":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Fgi Norm"}},"type":"object","required":["open_time"],"title":"FgiDailySchema"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PredictRequest":{"properties":{"symbol":{"type":"string","title":"Symbol","default":"XBX-USD"},"interval":{"type":"string","title":"Interval","default":"1d"},"horizon_days":{"type":"integer","maximum":7.0,"minimum":1.0,"title":"Horizon Days","default":1}},"type":"object","title":"PredictRequest"},"PredictResponse":{"properties":{"status":{"type":"string","title":"Status"},"cached":{"type":"boolean","title":"Cached"},"model_id":{"type":"string","title":"Model Id"},"as_of_time":{"type":"string","format":"date-time","title":"As Of Time"},"generated_at":{"type":"string","format":"date-time","title":"Generated At"},"valid_until":{"type":"string","format":"date-time","title":"Valid Until"},"horizon_days":{"type":"integer","title":"Horizon Days"},"predictions":{"items":{"$ref":"#/components/schemas/PredictionItem"},"type":"array","title":"Predictions"}},"type":"object","required":["status","cached","model_id","as_of_time","generated_at","valid_until","horizon_days","predictions"],"title":"PredictResponse"},"PredictionItem":{"properties":{"horizon_days":{"type":"integer","title":"Horizon Days"},"target_time":{"type":"string","format":"date-time","title":"Target Time"},"pred_open":{"type":"number","title":"Pred Open"},"pred_high":{"type":"number","title":"Pred High"},"pred_low":{"type":"number","title":"Pred Low"},"pred_close":{"type":"number","title":"Pred Close"},"pred_volume":{"type":"number","title":"Pred Volume"},"pred_components":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Pred Components"}},"type":"object","required":["horizon_days","target_time","pred_open","pred_high","pred_low","pred_close","pred_volume"],"title":"PredictionItem"},"TrainRequest":{"properties":{"symbol":{"type":"string","title":"Symbol","default":"XBX-USD"},"interval":{"type":"string","title":"Interval","default":"1d"},"feature_set":{"type":"string","title":"Feature Set","default":"full"},"lookback":{"type":"integer","maximum":365.0,"minimum":10.0,"title":"Lookback","default":60},"lr":{"type":"number","exclusiveMinimum":0.0,"title":"Lr","default":0.0001},"weight_decay":{"type":"number","minimum":0.0,"title":"Weight Decay","default":0.0005},"batch_size":{"type":"integer","maximum":4096.0,"minimum":8.0,"title":"Batch Size","default":256},"max_epochs":{"type":"integer","maximum":500.0,"minimum":1.0,"title":"Max Epochs","default":60},"min_epochs":{"type":"integer","maximum":500.0,"minimum":0.0,"title":"Min Epochs","default":10},"patience":{"type":"integer","maximum":200.0,"minimum":0.0,"title":"Patience","default":12},"min_delta":{"type":"number","minimum":0.0,"title":"Min Delta","default":0.0001},"seed":{"type":"integer","maximum":2147483647.0,"minimum":0.0,"title":"Seed","default":42},"holdout_from":{"type":"string","title":"Holdout From","default":"2025-06-01"}},"type":"object","title":"TrainRequest"},"TrainResponse":{"properties":{"status":{"type":"string","title":"Status"},"model_id":{"type":"string","title":"Model Id"},"trained_at":{"type":"string","format":"date-time","title":"Trained At"},"data_start":{"type":"string","format":"date-time","title":"Data Start"},"data_end":{"type":"string","format":"date-time","title":"Data End"},"metrics":{"additionalProperties":true,"type":"object","title":"Metrics"},"training_params":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Training Params"}},"type":"object","required":["status","model_id","trained_at","data_start","data_end","metrics"],"title":"TrainResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}